requirements.txt
# Primary requirements
-r ../pillar-python-sdk/requirements.txt

attrs==16.2.0
algoliasearch==1.12.0
bcrypt==3.1.3
blinker==1.4
bleach==1.4.3
celery[redis]==4.0.2
CommonMark==0.7.2
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
Eve==0.7.3
Flask==0.12
Flask-Babel==0.11.2
Flask-Cache==0.13.1
Flask-Script==2.0.6
Flask-Login==0.3.2
Flask-WTF==0.12
gcloud==0.12.0
google-apitools==0.4.11
httplib2==0.9.2
IPy==0.83
MarkupSafe==0.23
ndg-httpsclient==0.4.0
Pillow==4.1.1
python-dateutil==2.5.3
rauth==0.7.3
raven[flask]==6.3.0
redis==2.10.5
WebOb==1.5.0
wheel==0.29.0
zencoder==0.6.5

# Secondary requirements
amqp==2.1.4
billiard==3.5.0.2
Flask-PyMongo==0.4.1
-e git+git@github.com:armadillica/cerberus.git@sybren-0.9#egg=Cerberus
Events==0.2.2
future==0.15.2
html5lib==0.9999999
googleapis-common-protos==1.1.0
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.0.2
oauth2client==2.0.2
oauthlib==2.0.1
olefile==0.44
protobuf==3.0.0b2.post2
protorpc==0.11.1
pyasn1-modules==0.0.8
pymongo==3.4.0
pytz==2017.2
requests-oauthlib==0.7.0
rsa==3.4.2
simplejson==3.10.0
six==1.10.0
urllib3==1.22
vine==1.1.3
WTForms==2.1
Werkzeug==0.11.15
